Test scripts for REMOTE_LOCATION are setting the REMOTE_LOCATION env variable now via "process.env.REMOTE_LOCATION = '<value>';" and are forcing a reload of the proxy.js module. This avoids spawning two extra mocha processes Removed java hint from README.md Renamed remote_location to remoteLocation in proxy.js
