+++ lib/proxy.js
@@ -21,7 +21,7 @@ var env = {
 	noProxy: process.env.NO_PROXY || process.env.no_proxy,
 	httpProxy: process.env.HTTP_PROXY || process.env.http_proxy,
 	httpsProxy: process.env.HTTPS_PROXY || process.env.https_proxy,
-	remote_location: process.env.REMOTE_LOCATION || process.env.remote_location
+	remoteLocation: process.env.REMOTE_LOCATION || process.env.remote_location
 };
 
 // inspired by https://github.com/request/request/blob/33cd9e297a00c5540e55778a24a706effc35434c/request.js#L169
@@ -118,7 +118,7 @@ module.exports = function (options) {
 	var proxy = httpProxy.createProxyServer(options || {});
 
 	return function (req, res, next) {
-		var uri = createUri(req.url, env.remote_location);
+		var uri = createUri(req.url, env.remoteLocation);
 		if (!uri || !uri.host) {
 			return next();
 		}
@@ -148,3 +148,4 @@ module.exports = function (options) {
 //For testing
 module.exports._createUri = createUri;
 module.exports._buildRequestUrl = buildRequestUrl;
+module.exports._env = env;
